cicd-guardian-agent/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-multipart>=0.0.6

